Implement IMU calibration
Two services are added : calibrate_imu() and zet_zero().

calibrate_imu() computes the 3x3 scaling matrices and 3d bias vector for both
gyroscopes and accelerometers so that all data is returned in a consistent,
orthogonal frame of reference. This is done by implementing the paper "A robust
and easy to implement method for IMU calibration without external equipments,",
ICRA 2014. It requires no external sensor and a minimum of 10 static poses with
9 intervals with moderate motion in between.
The std dev of the sensor noise is also estimated and exported along with the
IMU data.

set_zero() computes a rotation matrix (merged into the current IMU calibration)
that aligns the gravity with the Z axis of the gyros/accelerometer in the
current robot configuration.