Raise an error in set_velocity only if a motor in in emergency state
Now that motors can be enabled while others are spinning, there is a transition
phase where a freshly enabled motor is not yet spinning while the others are
still receiving a velocity through set_velocity(). For this to work,
set_velocity() must not raise an error in this case.

Fixes #99