Get rid of superfluous internal state data after openrobots-idl upgrade
The rotorcraft interface of openrobots-idl>=1.4 now allows to directly export
all measured data, without creating an extra port. As a side effect, this
closes #110 since the energy_left in the rotor_measure port represents the
percentage of available battery.

While here, update the battery limits interface, so that a min/max battery
level can be specified. The energy_left member of the rotorcraft interface is
computed in % relative to those min/max limits.