Fix a logic bug in the startup monitor codel
If a motor with a low id would be slow to start, a motor with a higher id could
already have gone through the whole init->start->failure loop and would then
appear as not yet started. This would cause the 'start' activity hand forever.

To prevent this, make sure to scan all motors in every loop of the monitor
state of the start activity instead of returning the new activity state as soon
as it is known.