Use aio(7) for logging and drop log entries when io requests are too slow
Use aio_write(3) to append to the log files. If a previous aio request was
not completed after one period, drop the next log entry and add an empty
line in the log file to indicate the miss (gnuplot will show it nicely).

Also add a log_info() function reporting the number of missed log entries.