Reset the time offset of the received data if it diverges too much
The timestamp estimation assumes that the sensors (IMU, motors) are streaming
data at the specified frequency. It can happen that this is not true in
simulation (or if the flight controller did not receive the streaming message).

To avoid to articifially publish data too much in the past, reset the time
offset if the delay is greater than 2 periods.