Rename the component to "rotorcraft"
This component will control newer paparazzi-based hardware, so the name
'mikrokopter' would be a misnomer.

In addition, the mikrokopter flight controller 2.5 hardware is obsolete and
bound to die sooner or later. Newer mikrokopter flight controller hardware will
unlikely be supported as it lack documentation to do so, and the paparazzi
hardware is much better anyway. But the old hardware is still supported.

Bump version to 3.0