Check sensor rate after motor startup and add a configurable timeout
The startup phase may send repeated 'motor start' command to the hardware
when a motor does not start immediately. When the repeat rate is too fast,
this prevents the motor data from beeing sent by the motor and makes the sensor
rate drop temporarily.

Since the 'servo' service checks the sensor rate, it might stop early if called
right after the start, when the rate has dropped too much.

To prevent this:
 * Don't send repeated start command too fast (sensor rate drops less)
 * Wait after all motor have started that the sensor rate is above the
   threshold.
 * Add a startup timeout to prevent blocking the the start phase if for some
   reason the sensor rate cannot be achieved.