Update data reception rate estimator
Instead of a simple low-pass filter that is not robust to packet losses, use an
adaptive gain filter which rejects outliers well.

The servo() activity will now stop if the actual rate is below 90% of the
target. It will also stop if there are ten packets loss in a row.