FIx expected norm of the magnetic field during magnetometer calibration
Because of the bias, the expected norm of the magnetic field cannot be just the
average of the raw samples norm. Estimate it using the min/max average of raw
samples (as with the previous simple calibration).

This is not critical as the norm is in practice not used by the attitude
estimation (only the direction). But better have something realistic anyway.