Update IMU resolution messages on the chimera
The chimera flight controller now uses a different resolution for sending
accelerometer and gyroscope data, so as to avoid saturations.
The messages are sent with native resolution: 8g on 15 bits for accelerometer
and 1000deg/s on 15 bits for gyroscope.

While here, factor out settings specific to flight controller boards in a
common structure.

Require chimera-1.1 firmware revision.