Add an optional log file for IMU calibration
If specified, the log file will be filled with calibrated values of the IMU
used for calibration. This is useful to check for any acquisition artifacts
that might have altered the results.

The average acceleration and angular velocities (during standstill intervals)
are also displayed after calibration. They are supposed to be 9.81m/s² and
0rad/s respectively.