Add a setter for the tolerance of the standstill detector in calibration
The settings of the standstill detector might have been a bit loose, so
thighten it. In order to not create trouble, make this configurable with
set_calibration_param(). Log the detector results in the log file.

While here, include gyroscope data in the standstill detector. Although
unlikely, it is still possible in practice to have the IMU rotating
with no acceleration.