Don't fail servo() when motors are starting
The safety check in servo() that checks if motors are spinning is too strong
when propellers spinning direction must be reverted (see issue #336).

Since the hardware will ignore any velocity setpoint until motors are properly
started, it is safe to weaken the safety check in servo() by just checking that
motors are starting or spinning.

However, the initial acceleration ramp must be started only when motors are
fully spinning.