Avoid numerical unstability during sample acquisition for calibration
The motion detection used in calibration is based on the global minimum
variance of each sensor. If, for some reason, there is not enough noise, the
minimum variance will go very low and lead to numerical unstability.

Use an absolute minimum variance of 1e-4 to avoid that. This should help with
issue #339.